+++ CHANGELOG.md
@@ -16,9 +16,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Observers `on_push_frame()` now take a single argument `FramePushed` instead
+  of multiple arguments.
+
 - Updated the default voice for `DeepgramTTSService` to `aura-2-helena-en`.
 
+### Deprecated
+
+- Observer `on_push_frame(src, dst, frame, direction, timestamp)` is now
+  deprecated, use `on_push_frame(data: FramePushed)` instead.
+
 ### Fixed
+
+- Fixed a `UltravoxSTTService` issue that would cause the service to generate
+  all tokens as one word.
 
 - Fixed a `PipelineTask` issue that would cause tasks to not be cancelled if
   task was cancelled from outside of Pipecat.